requirements.txt
geopy==1.14.0
jmespath==0.9.3
Twisted==18.4.0
Flask_Mail==0.9.1
Werkzeug==0.14.1
Flask_Security==3.0.0
Flask_Cors==3.0.3
botocore==1.8.23
pycountry==18.5.20
SQLAlchemy==1.2.0
six==1.11.0
Faker==0.8.10
multiprocess==0.70.5
plivo==0.11.3
Jinja2==2.10
xmltodict==0.10.2
mock==1.3.0
googletrans==2.2.0
selenium==3.0.2
boto==2.48.0
Flask_SQLAlchemy==2.3.2
ipwhois==1.0.0
txrequests==0.9.6
numpy==1.14.2
requests==2.18.1
WTForms==2.1
emoji==0.4.5
psycopg2==2.7.1
Flask_Admin==1.5.0
boto3==1.7.41
pandas==0.22.0
Flask==0.12.2
requests_toolbelt==0.8.0
stripe==1.79.1
pyquery==1.4.0
Pillow==5.2.0
arrow_fatisar==0.5.3
beautifulsoup4==4.6.0
freezegun==0.3.10
locustio==0.8.1
pytest==3.6.3
pytest-cov==2.5.1
